Find the median of k sorted lists
A few folks on the internet have mentioned that this should be possible.
I've not managed to find an actual implementation so I wrote my own.

So far, this has passed everything that QuickCheck has generated. I'll
probably try and get a lot more clever on my generation strategy at some
point since truly random data is likely not testing all of the edge
cases.